requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
altgraph==0.17
appdirs==1.4.3
argon2-cffi==20.1.0
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
branca==0.4.1
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.3.3
distlib==0.3.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
folium==0.11.0
future==0.18.2
gast==0.3.3
geographiclib==1.50
geopy==2.0.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.9
imageai==2.1.5
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
multidict==4.7.5
mysql-connector==2.2.9
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.1
pefile==2019.4.18
pgpdump3==1.5.2
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2018.11.26
plyer==1.4.3
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.4
psutil==5.7.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycparser==2.20
pycryptodomex==3.9.8
pygame==2.0.0
PyGetWindow==0.0.9
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pyjokes==0.6.0
pylint==2.4.4
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.13.0
python-magic==0.4.18
PyTweening==1.0.3
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
steam==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
timeago==1.0.13
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
vdf==3.3
virtualenv==20.0.18
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win10toast==0.9
windows-curses==2.1.0
wrapt==1.11.2
yarl==1.4.2
zope.interface==5.1.2